web/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
alembic==1.16.5
allure-pytest==2.15.3
argon2-cffi==25.1.0
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
defusedxml==0.7.1
distro==1.9.0
dotenv==0.9.9
fastapi==0.134.0
fonttools==4.61.1
Faker==40.8.0
factory-boy==3.3.3
fpdf2==2.8.7
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-vision==3.12.1
google-genai==1.65.0
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
hypothesis==6.151.9
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
librt==0.8.1
locust==2.43.3
MarkupSafe==3.0.3
mypy==1.19.1
mypy_extensions==1.1.0
pathspec==1.0.4
pillow==12.1.1
playwright==1.58.0
pre-commit==4.3.0
proto-plus==1.27.1
protobuf==5.29.6
pypdf==6.7.5
qrcode==8.2
pyee==13.0.1
pg8000==1.31.5
psycopg[binary]==3.3.3
psycopg2-binary==2.9.11
pytest==9.0.2
pytest-base-url==2.1.0
pytest-cov==7.0.0
pytest-html==4.2.0
pytest-playwright==0.7.2
pytest-randomly==4.0.1
pytest-timeout==2.4.0
pytest-xdist==3.8.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-sdk==1.38.0
pwdlib==0.3.0
pyparsing==3.3.2
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
python-slugify==8.0.4
requests==2.32.5
redis==7.3.0
rsa==4.9.1
ruff==0.13.2
schemathesis==4.12.0
sentry-sdk[fastapi]==2.31.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.47
starlette==0.52.1
tenacity==9.1.4
text-unidecode==1.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
websockets==16.0
